Moved `storage.js` to `utils/storage.ts`
