+++ src/popup/popup.js
@@ -8,7 +8,7 @@ import { disable, enable, hide, show, click } from "../utils/html";
 import { setupReleasesTab } from "./tabs/releases_tab.js";
 import { setupReleaseTab } from "./tabs/release_tab.js";
 import { setupCsvDataTab } from "./tabs/csv_data_tab.js";
-import { getStorageSize } from "../modules/storage.js";
+import { getStorageSize } from "../utils/storage.js";
 import { bytesToSize } from "../utils/utils";
 import { setupConsole, setupConsoleRelease } from "./console.js";
 import { isValidBandcampURL } from "../bandcamp/modules/html.js";
