+++ src/popup/tabs/releases_tab.js
@@ -1,7 +1,7 @@
 import { ReleaseItem } from "../../app/release.js";
 import { getCurrentTab, openTabs } from "../../utils/chrome";
 import { input } from "../../utils/html";
-import { findReleasesByUrls } from "../../modules/storage.js";
+import { findReleasesByUrls } from "../../utils/storage.js";
 import { populateReleasesList, removeButtonLoadingState, setBackgroundImage, setButtonInLoadingState } from "../helpers.js";
 import { downloadReleasesCsv } from "./download_tab.js";
