+++ src/utils/storage.ts
@@ -1,7 +1,7 @@
 import { Release } from "../app/release.js";
 import { isValidBandcampURL } from "../bandcamp/modules/html.js";
-import { generateKeyForRelease, generateKeyForUrl, generateKeyUrlMapFromUrls, generateKeysFromUrls } from "../utils/key-generator";
-import { isArray, isFunction, isObject } from "../utils/utils";
+import { generateKeyForRelease, generateKeyForUrl, generateKeyUrlMapFromUrls, generateKeysFromUrls } from "./key-generator.js";
+import { isArray, isFunction, isObject } from "./utils.js";
 
 const storage = chrome.storage.local;
